requirements.txt
streamlit
numpy
pandas
torch
torchvision
pytorch_lightning
torchmetrics
torchsummary
Pillow